app/requirements.txt
fastapi
uvicorn[standard]
celery
redis
python-docx
